requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.6.15
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx>=0.25.2,<1
idna==3.10
jiter==0.10.0
openai==1.93.0
pip==24.2
pydantic==2.11.7
pydantic-core==2.33.2
python-dotenv==1.1.1
python-telegram-bot==20.7
nest_asyncio==1.6.0
sniffio==1.3.1
tqdm==4.67.1
typing-extensions==4.14.0
typing-inspection==0.4.1
